+++ src/frontend/app/(time)/time/[year]/m/[month]/page.tsx
@@ -1,16 +1,33 @@
-import { TimePage } from '@/app/(pages)/time/time';
+import { TimePage } from '@/app/(time)/time/time';
 import {
 	fetchBoxOfficeInfinite,
 	fetchBoxOfficeSummary,
 } from '@/lib/api/dataFetching';
+import { getLastDayofMonth } from '@/lib/helpers/dates';
 
 export async function generateMetadata({
 	params,
 }: {
-	params: { year: number };
+	params: { year: string; month: string };
 }) {
-	const title = `${params.year} | Box Office Data`;
-	const description = `${params.year} | Box Office Data`;
+	const months = [
+		'January',
+		'February',
+		'March',
+		'April',
+		'May',
+		'June',
+		'July',
+		'August',
+		'September',
+		'October',
+		'November',
+		'December',
+	];
+	const m = parseInt(params.month);
+
+	const title = `${months[m - 1]} ${params.year} | Box Office Data`;
+	const description = `${months[m - 1]} ${params.year} | Box Office Data`;
 
 	return {
 		title: title,
@@ -43,20 +60,32 @@ export default async function Page({
 	params,
 	searchParams,
 }: {
-	params: { year: string };
+	params: { year: string; month: string };
 	searchParams: { country: string; distributor: string };
 }) {
-	// Build Dates based on existing params or defaults.
-	const start = new Date(parseInt(params.year), 0, 1);
 	const countries = searchParams?.country?.split(',').map(Number);
 	const distributors = searchParams?.distributor?.split(',').map(Number);
+	// Build Dates based on existing params or defaults.
+	const start = new Date(parseInt(params.year), parseInt(params.month) - 1, 1);
 
-	// Check if the passed year is the current year
+	// Check if the passed year and month are the current year and current month
 	const currentYear = new Date().getFullYear();
-	const end =
-		parseInt(params.year) === currentYear
-			? new Date()
-			: new Date(parseInt(params.year), 11, 31);
+	const currentMonth = new Date().getMonth() + 1;
+	const isCurrentYear = parseInt(params.year) === currentYear;
+	const isCurrentMonth =
+		isCurrentYear && parseInt(params.month) === currentMonth;
+
+	// Adjust the end date based on whether it's the current year and current month
+	let end: Date;
+	if (isCurrentMonth) {
+		end = new Date(); // Set the end date to today if it's the current month
+	} else {
+		end = new Date(
+			parseInt(params.year),
+			parseInt(params.month) - 1,
+			getLastDayofMonth(parseInt(params.month))
+		); // Set the end date to the last day of the specified month
+	}
 
 	// Build Date Strings for API
 	const startDate = `${start.getFullYear()}-${
@@ -82,6 +111,7 @@ export default async function Page({
 	return (
 		<TimePage
 			year={parseInt(params.year)}
+			month={parseInt(params.month)}
 			results={results}
 			timeComparisonData={timeComparisonData.results}
 		/>
