+++ src/frontend/app/(time)/time/[year]/page.tsx
@@ -1,37 +1,16 @@
-import { TimePage } from '@/app/(pages)/time/time';
+import { TimePage } from '@/app/(time)/time/time';
 import {
 	fetchBoxOfficeInfinite,
 	fetchBoxOfficeSummary,
 } from '@/lib/api/dataFetching';
-import addDays from 'date-fns/addDays';
 
 export async function generateMetadata({
 	params,
 }: {
-	params: { year: number; month: string; day: number };
+	params: { year: number };
 }) {
-	const months = [
-		'January',
-		'February',
-		'March',
-		'April',
-		'May',
-		'June',
-		'July',
-		'August',
-		'September',
-		'October',
-		'November',
-		'December',
-	];
-	const m = parseInt(params.month);
-
-	const title = `${params.day} ${months[m - 1]} ${
-		params.year
-	} | Box Office Data`;
-	const description = `${params.day} ${months[m - 1]} ${
-		params.year
-	} | Box Office Data`;
+	const title = `${params.year} | Box Office Data`;
+	const description = `${params.year} | Box Office Data`;
 
 	return {
 		title: title,
@@ -64,23 +43,26 @@ export default async function Page({
 	params,
 	searchParams,
 }: {
-	params: { year: string; month: number; day: number };
+	params: { year: string };
 	searchParams: { country: string; distributor: string };
 }) {
+	// Build Dates based on existing params or defaults.
+	const start = new Date(parseInt(params.year), 0, 1);
 	const countries = searchParams?.country?.split(',').map(Number);
 	const distributors = searchParams?.distributor?.split(',').map(Number);
-	// Build Dates based on existing params or defaults.
-	const start = new Date(parseInt(params.year), params.month - 1, params.day);
-	const sLastWeek = addDays(start, -7);
+
+	// Check if the passed year is the current year
+	const currentYear = new Date().getFullYear();
+	const end =
+		parseInt(params.year) === currentYear
+			? new Date()
+			: new Date(parseInt(params.year), 11, 31);
 
 	// Build Date Strings for API
 	const startDate = `${start.getFullYear()}-${
 		start.getMonth() + 1
 	}-${start.getDate()}`;
-
-	const startLastWeek = `${sLastWeek.getFullYear()}-${
-		sLastWeek.getMonth() + 1
-	}-${sLastWeek.getDate()}`;
+	const endDate = `${end.getFullYear()}-${end.getMonth() + 1}-${end.getDate()}`;
 
 	// Fetch data
 	let yearsToGoBack = 25;
@@ -90,26 +72,17 @@ export default async function Page({
 	}
 
 	const { results, isReachedEnd, percentFetched } =
-		await fetchBoxOfficeInfinite(startDate, startDate, distributors, countries);
-	const { results: lastWeekResults } = await fetchBoxOfficeInfinite(
-		startLastWeek,
-		startLastWeek,
-		distributors,
-		countries
-	);
+		await fetchBoxOfficeInfinite(startDate, endDate, distributors, countries);
 	const timeComparisonData = await fetchBoxOfficeSummary(
 		startDate,
-		startDate,
+		endDate,
 		yearsToGoBack
 	);
 
 	return (
 		<TimePage
 			year={parseInt(params.year)}
-			month={params.month}
-			day={params.day}
 			results={results}
-			lastWeekResults={lastWeekResults}
 			timeComparisonData={timeComparisonData.results}
 		/>
 	);
