+++ src/frontend/app/(time)/time/layout.tsx
@@ -1,0 +1,33 @@
+import { ControlsWrapper } from '@/components/controls';
+import { CountryFilter } from '@/components/country-filter';
+import { BreadcrumbsTime } from '@/components/custom/breadcrumbs-time';
+
+import { mapToValues } from '@/lib/helpers/filters';
+import { fetchCountryList } from '@/lib/api/dataFetching';
+import { Suspense } from 'react';
+import { TimePage } from './time';
+
+export default async function Layout({
+	children,
+}: {
+	children: React.ReactNode;
+}) {
+	const countryData = await fetchCountryList(1, 100);
+	const countryOptions = mapToValues(countryData.results);
+
+	return (
+		<>
+			<ControlsWrapper className='hidden md:flex'>
+				<BreadcrumbsTime />
+				{/* <Suspense
+					fallback={
+						<TimePage year={2023} results={[]} timeComparisonData={[]} />
+					}
+				> */}
+				<CountryFilter countries={countryOptions} />
+				{/* </Suspense> */}
+			</ControlsWrapper>
+			{children}
+		</>
+	);
+}
