+++ src/frontend/app/loading.tsx
@@ -1,5 +1,6 @@
 import { Skeleton } from '@/components/skeleton';
+import { Spinner } from '../components/custom/spinner';
 
 export default function Loading() {
-	return <Skeleton />;
+	return <Spinner />;
 }
